Rename benchmark_utils to benchmark_util to adhere to existing naming patterns
